Move StatusComponent in separate file
